Split authorised API bits into a separate module to copy structure of the IdeaModule.
