+++ src/Ideastrike.Nancy/Ideastrike.Nancy.csproj
@@ -190,6 +190,7 @@
     <Compile Include="Models\Vote.cs" />
     <Compile Include="Modules\AdminModule.cs" />
     <Compile Include="Modules\ApiModule.cs" />
+    <Compile Include="Modules\ApiModule.Secured.cs" />
     <Compile Include="Modules\CommentModule.cs" />
     <Compile Include="Modules\FeatureModule.cs" />
     <Compile Include="Modules\IdeaModule.Secured.cs" />
